Split erb.rb into files for each module under ERB
